core: fix allocate memory once for *GROUPMEMBERS macros (Stephane Lapie) #1076
[--SNIP--]
Previously we used to iterate over all available members once and
issue a realloc for every member. Such memory thrashing provides
very poor performance compared to just counting the length we need
and allocate it properly once and then writing to the buffer thus
allocated.

This patch corrects that for host and servicegroupmembers.

Signed-off-by: Stephane Lapie <stephane.lapie@darkbsd.org>
Signed-off-by: Andreas Ericsson <ae@op5.se>

Author: Stephane Lapie <stephane.lapie@darkbsd.org>
[--SNIP--]

fixes #1076