core: fix race condition on flexible downtime commands when duration not set or zero #1061
it may happen that an external command sets a
flexible downtime with duration not set or zero.

a flexible downtime is triggered between start and
end time and needs a duration to be set and non zero.

without the fix, the scheduled downtime starts and then
tries to calculate the end time out of the set duration
for the downtime and scheduling a new event.
the time passed over there is in the past, and causes the
core to get into a loop.

to prevent this misbehavior, a duration of null or zero is
not allowed and triggered as error in base/commands.c not
setting any flexible downtime.

thanks to Daniel Heule for reporting the issue and testing
my proposed fix ok.

fixes #1061